requirements.txt
faster-whisper
pysubs2
python-dotenv
google-generativeai
flask
flask-socketio
